Implement login form component
This commit introduces the Login component in the application,
providing a form for users to enter their email and password for authentication.
The form leverages the Redux 'useDispatch' and 'useSelector' hooks to manage state.

Key features and functions included are:

1. State management for 'email' and 'password' fields.
2. Redux actions for login and clear error/success messages.
3. Utilization of 'useEffect' and 'useCallback' hooks to manage side effects related to authentication status and navigation.
4. Conditional redirection to the dashboard or the previous page after successful login, managed with the 'useNavigate' and 'useLocation' hooks from react-router-dom.
5. An inclusion of the StatusMessage component to display authentication error/success messages.
6. Styling and structure of the form for a better user experience.
7. An option to navigate to the registration page if a user does not have an account.

This enhancement ensures a secure and efficient way for users to authenticate themselves in the application.